requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==1.4.23
psycopg2==2.9.9
python-multipart==0.0.5
python-jose[cryptography]==3.2.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pillow==10.0.1
numpy==1.24.3
requests==2.31.0
qrcode[pil]==7.4.2
reportlab==4.0.8
python-dotenv==1.0.0
twilio==8.10.3
pgvector==0.2.4
# gradio_client==1.3.0  # Removed - using local model
fastapi-mail==1.4.1
pydantic==2.5.0
# PyTorch for local model inference
torch==2.0.1
torchvision==0.15.2
# opencv-python-headless==4.8.1.78